# docs/openapi/tenant-licensing.yaml
openapi: 3.0.3
info:
  title: Maniforge Tenant/Licensing API
  version: 0.2.0
  description: |
    Contract for tenant lifecycle, licensing, managed-tenant grants, entitlements, quota,
    lifecycle events, and internal access-state / event delivery.
servers:
  - url: /tenant-licensing
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AccessState:
      type: object
      required: [ok, tenant_code, subtenant_code, tenant_active, subtenant_active, license_active]
      properties:
        ok: { type: boolean, enum: [true] }
        tenant_code: { type: string }
        subtenant_code: { type: string }
        tenant_active: { type: boolean }
        subtenant_active: { type: boolean }
        license_active: { type: boolean }
        features:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        checked_at: { type: string }
security:
  - bearerAuth: []
paths:
...
  /api/v1/tenants:
    get:
      responses:
        '200':
          description: Tenant list.
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code: { type: string }
                name: { type: string }
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Tenant created.
  /api/v1/tenants/{tenantCode}:
    patch:
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                status: { type: string, enum: [active, suspended, disabled] }
      responses:
        '200':
          description: Tenant updated.
...
  /api/v1/tenants/{tenantCode}/subtenants/{subtenantCode}:
    patch:
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema: { type: string }
        - name: subtenantCode
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                status: { type: string, enum: [active, suspended, disabled] }
      responses:
        '200':
          description: Subtenant updated.
  /api/v1/plans:
    get:
      responses:
        '200':
          description: License plans.
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code: { type: string }
                name: { type: string }
                status: { type: string, enum: [active, disabled] }
                features:
                  type: object
                  additionalProperties: true
                limits:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: License plan created or updated.
...
  /api/v1/tenants/{tenantCode}/entitlements:
    get:
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tenant entitlements.
...
  /api/v1/events:
    get:
      parameters:
        - name: tenant_code
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
      responses:
        '200':
          description: Tenant lifecycle events.
  /internal/v1/tenants/{tenantCode}/subtenants/{subtenantCode}/access-state:
    get:
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema: { type: string }
        - name: subtenantCode
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Runtime access state for RBAC.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessState'
...
  /internal/v1/events/{eventId}/ack:
    post:
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Event acknowledged.